create the server atomicStack

+++ runtime/gc/collector/ipc_server_sweep.h
@@ -33,7 +33,7 @@ class IPCServerMarkerSweep {
 
   gcservice::GCServiceClientRecord* const client_rec_;
   space::GCSrvSharableHeapData* const heap_meta_;
-  const uintptr_t offset_;
+  const uint32_t offset_;
 
   GCSrverCollectorSpace spaces_[KGCSpaceCount];
   space::GCSrvSharableCollectorData* curr_collector_ptr_;
